pyproject.toml
[project]
name = "pomodoro"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexandr Novoseletsky",email = "night2324@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "types-python-jose (>=3.5.0.20250531,<4.0.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.5,<3.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "ruff (>=0.14.4,<0.15.0)",
    "sqlalchemy2-stubs (>=0.0.2a38,<0.0.3)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "webptools (>=0.0.9,<0.0.10)",
    "pillow (>=12.0.0,<13.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "black (>=25.11.0,<26.0.0)",
    "pydocstringformatter (>=0.7.5,<0.8.0)",
]

[tool.poetry]
packages = [{include = "pomodoro", from = "."}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

[tool.ruff]
line-length = 79
fix = true

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
    "migrations",
    "alembic",
    "scripts",
]

[tool.ruff.format]
docstring-code-line-length = 72
docstring-code-format = true
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
    "D",    # pydocstyle
]
ignore = [
    "D203",  # или "D211", если хочешь наоборот
    "D213",  # или "D212", если хочешь второй стиль
    "RUF002",  # Игнорирование кирилических символов
    "RUF001",  # Игнорирование кирилических символов
    "RUF003",  # Игнорирование кирилических символов
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true


[tool.pydocstringformatter]
write = true
max-line-length = 72
linewrap-full-docstring = true
split-summary-body = false
strip-whitespaces = true
style = ["pep257"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"